refactor:separete to small components
