+++ src/components/OrderTable/OrderTable.tsx
@@ -16,8 +16,8 @@ import { useAppState } from '../../context/AppState';
 import useLocalStorage from '../../hooks/useLocalStorage';
 import useRealTimeData from '../../hooks/useRealTimeData';
 import NoData from '../../ui/NoData';
-import ChipColor from '../ChipColor';
-import { Status } from '../ChipColor/ChipColor';
+import ChipColor from '../../ui/ChipColor';
+import { Status } from '../../ui/ChipColor/ChipColor';
 import Pagination from '../Pagination';
 import RowMenu from '../RowMenu';
 import TableFilters from '../TableFilters';
