+++ src/ui/Sidebar/Sidebar.tsx
@@ -1,54 +1,30 @@
-import Avatar from '@mui/joy/Avatar';
-import Box from '@mui/joy/Box';
-import Divider from '@mui/joy/Divider';
-import IconButton from '@mui/joy/IconButton';
-import Input from '@mui/joy/Input';
-import List from '@mui/joy/List';
-import ListItem from '@mui/joy/ListItem';
-import ListItemButton, { listItemButtonClasses } from '@mui/joy/ListItemButton';
-import Typography from '@mui/joy/Typography';
-import Drawer from '@mui/joy/Drawer';
 import SearchRoundedIcon from '@mui/icons-material/SearchRounded';
-import DescriptionIcon from '@mui/icons-material/Description';
+import SettingsRoundedIcon from '@mui/icons-material/SettingsRounded';
 import ShoppingCartRoundedIcon from '@mui/icons-material/ShoppingCartRounded';
 import SupportRoundedIcon from '@mui/icons-material/SupportRounded';
-import SettingsRoundedIcon from '@mui/icons-material/SettingsRounded';
-import LogoutRoundedIcon from '@mui/icons-material/LogoutRounded';
-import ColorSchemeToggle from '../ColorSchemeToggle';
+// import DescriptionIcon from '@mui/icons-material/Description'; TO DO: uncomment after add new feature
+import Box from '@mui/joy/Box';
+import Divider from '@mui/joy/Divider';
+import Drawer from '@mui/joy/Drawer';
+import Input from '@mui/joy/Input';
+import List from '@mui/joy/List';
+import { listItemButtonClasses } from '@mui/joy/ListItemButton';
 import ModalClose from '@mui/joy/ModalClose';
-
-import { Card, CardContent, Sheet } from '@mui/joy';
+import ColorSchemeToggle from '../ColorSchemeToggle';
+import { Sheet } from '@mui/joy';
 import { useAppState } from '../../context/AppState';
-import { signOut } from 'firebase/auth';
-import { useSnackbar } from 'notistack';
-import { useNavigate } from 'react-router-dom';
-import { auth } from '../../api/firebase';
-import RoutesEnum from '../../types/routesEnum';
 import { useMediaQuery } from '@mui/material';
+import RoutesEnum from '../../types/routesEnum';
 import LanguageToggle from '../LanguageToggle';
+import UserAvatar from '../UserAvatar';
+import MenuCards from '../MenuCards';
+import MenuItems from '../MenuItems';
 
 export default function Sidebar(): JSX.Element {
   const {
     state: { showSidebar },
     dispatch,
   } = useAppState();
-
-  const navigate = useNavigate();
-
-  const { enqueueSnackbar } = useSnackbar();
-
-  const logOut = async () => {
-    await signOut(auth)
-      .then(() => {
-        navigate(RoutesEnum.LOGIN);
-        dispatch({ type: 'SET_REGISTER_APP', payload: false });
-        dispatch({ type: 'SET_SHOW_SIDEBAR', payload: false });
-        enqueueSnackbar('Logout completed successfully. See you next time!', { variant: 'success' });
-      })
-      .catch((error) => {
-        enqueueSnackbar(error.message, { variant: 'error' });
-      });
-  };
 
   const isMobile = useMediaQuery('(max-width:600px)');
 
@@ -58,11 +34,12 @@ export default function Sidebar(): JSX.Element {
       icon: <ShoppingCartRoundedIcon color='success' />,
       redirect: RoutesEnum.ORDERS,
     },
-    {
-      label: 'Invoice',
-      icon: <DescriptionIcon color='primary' />,
-      redirect: RoutesEnum.INVOICE,
-    },
+    // TO DO: Add new feature
+    // {
+    //   label: 'Invoice',
+    //   icon: <DescriptionIcon color='primary' />,
+    //   redirect: RoutesEnum.INVOICE,
+    // },
   ];
 
   const itemsAdmin = [
@@ -141,23 +118,7 @@ export default function Sidebar(): JSX.Element {
             }}
           >
             {items.map(({ label, icon, redirect }) => (
-              <Card
-                key={label}
-                sx={{
-                  boxShadow: 'none',
-                  m: 0.6,
-                  display: 'flex',
-                  alignItems: 'center',
-                  flexDirection: 'row',
-                  '&:hover': { bgcolor: 'background.level1', border: '1px solid green' },
-                }}
-                onClick={() => navigate(redirect)}
-              >
-                <Box>{icon}</Box>
-                <CardContent>
-                  <Typography level='title-md'>{label}</Typography>
-                </CardContent>
-              </Card>
+              <MenuCards label={label} icon={icon} redirect={redirect} />
             ))}
 
             <List
@@ -170,20 +131,8 @@ export default function Sidebar(): JSX.Element {
                 mb: 2,
               }}
             >
-              {itemsAdmin.map(({ label, icon }) => (
-                <ListItem
-                  sx={{
-                    '&:hover': {
-                      border: '1px solid green',
-                    },
-                  }}
-                  key={label}
-                >
-                  <ListItemButton disabled={true}>
-                    {icon}
-                    {label}
-                  </ListItemButton>
-                </ListItem>
+              {itemsAdmin.map(({ label, icon, redirect }) => (
+                <MenuItems label={label} icon={icon} redirect={redirect} />
               ))}
             </List>
           </Box>
@@ -194,16 +143,7 @@ export default function Sidebar(): JSX.Element {
             }}
           />
 
-          <Box sx={{ display: 'flex', gap: 1, alignItems: 'center', m: 2 }}>
-            <Avatar variant='outlined' size='sm' src='src/assets/avatars.png' />
-            <Box sx={{ minWidth: 0, flex: 1 }}>
-              <Typography level='title-sm'>Jon Do.</Typography>
-              <Typography level='body-xs'>jon@do.com</Typography>
-            </Box>
-            <IconButton size='sm' variant='plain' color='danger' onClick={logOut}>
-              <LogoutRoundedIcon />
-            </IconButton>
-          </Box>
+          <UserAvatar />
         </Drawer>
       </Sheet>
     </>
